Sometimes when carbon server initialization process try to get an public ip in a loop, it is abruptly stopped by the linux initialization and instance is created without public ip retrieved and without java process started. Fixed this
git-svn-id: https://svn.wso2.org/repos/wso2/carbon/platform/trunk@135239 a5903396-d722-0410-b921-86c7d4935375

+++ build/tropos/image_create/init_scripts/carbon/wso2-openstack-init.sh
@@ -10,11 +10,12 @@ export REPO_PATH_EBS=""
 export CONF_PATH_EBS=""
 export PRODUCT_NAME=""
 export CONTROLLER_IP=""
+CRON_DURATION=1
+PUBLIC_IP=""
 
 
 #Check whether if any java processes are running
-
-if [ "$(pidof java)" ]; then
+if [[ "$(pidof java)" ]]; then
    # process was found
    echo "An already running java process is found. Exiting..." >> $LOG
    exit 0;
@@ -25,56 +26,78 @@ fi
 
 echo ---------------------------- >> $LOG
 
-# payload will be copied into ${instance_path}/user-data.txt file
-# that file will be renamed as zip file and extract, 
+if [ ! -d ${instance_path}/payload ]; then
+    # payload will be copied into ${instance_path}/user-data.txt file
+    # that file will be renamed as zip file and extract, 
 
-# copy user-data.txt and rename as payload.zip file
-cp ${instance_path}/user-data.txt ${instance_path}/payload.zip
-unzip ${instance_path}/payload.zip -d ${instance_path}/
+    # copy user-data.txt and rename as payload.zip file
+    cp ${instance_path}/user-data.txt ${instance_path}/payload.zip
+    unzip ${instance_path}/payload.zip -d ${instance_path}/
 
-# if error code is 0, there was no error
-if [ "$?" = "0" ]; then
-	echo retrieved data >> $LOG
-	rm -Rf ${instance_path}/payload
-	unzip ${instance_path}/payload.zip -d ${instance_path}/
-	# if unzip error code is 0, there was no error
-	if [ "$?" = "0" ]; then
-		echo Extracted payload >> $LOG
-	else
-		echo rc.local : payload.zip is corrupted >> $LOG
-	fi
+    # if error code is 0, there was no error
+    if [ "$?" = "0" ]; then
+        echo retrieved data >> $LOG
+        rm -Rf ${instance_path}/payload
+        unzip ${instance_path}/payload.zip -d ${instance_path}/
+        # if unzip error code is 0, there was no error
+        if [ "$?" = "0" ]; then
+            echo Extracted payload >> $LOG
+        else
+            echo rc.local : payload.zip is corrupted >> $LOG
+        fi
 
-else
-	echo rc.local : error retrieving user data >> $LOG
+    else
+        echo rc.local : error retrieving user data >> $LOG
+    fi
+
+    chmod -R 0600 ${instance_path}/payload/wso2-key
+    cat ${instance_path}/payload/known_hosts >> ~/.ssh/known_hosts
+
+    # Write a cronjob to execute wso2-openstack-init.sh periodically until public ip is assigned
+    crontab -l > ./mycron
+    echo "*/${CRON_DURATION} * * * * /opt/wso2-openstack-init.sh > /var/log/wso2-openstack-init.log" >> ./mycron
+    crontab ./mycron
+    rm ./mycron
+
+    for i in `/usr/bin/ruby /opt/get-launch-params.rb`
+    do
+        echo "export" ${i} >> /home/ubuntu/.bashrc
+    done
+
 fi
 
-chmod -R 0600 ${instance_path}/payload/wso2-key
-cat ${instance_path}/payload/known_hosts >> ~/.ssh/known_hosts
 
+wget http://169.254.169.254/latest/meta-data/public-ipv4
+files="`cat public-ipv4`"
+if [[ -z ${files} ]]; then
+    echo "getting public ip" >> $LOG
+    for i in {1..30}
+    do
+      rm -f ./public-ipv4
+      wget http://169.254.169.254/latest/meta-data/public-ipv4
+      files="`cat public-ipv4`"
+      if [ -z $files ]; then
+          echo "Public ip is not yet assigned. Wait and continue for $i the time ..." >> $LOG
+          sleep 1
+      else
+          echo "Public ip assigned" >> $LOG
+          break
+      fi
+    done
 
-echo "getting public ip" >> $LOG
-while [ 1 > 0 ]
-do
-  rm -f ./public-ipv4
-  wget http://169.254.169.254/latest/meta-data/public-ipv4
-  files="`cat public-ipv4`"
-  if [ -z $files ]; then
-      echo "Public ip is not yet assigned. Wait and continue ..." >> $LOG
-      sleep 1
-      continue
-  else
-      echo "Public ip assigned" >> $LOG
-      break
-  fi
-done
+    if [ -z $files ]; then
+      echo "Public ip is not yet assigned. Exiting ..." >> $LOG
+      exit 0
+    fi
+    for x in $files
+    do
+        PUBLIC_IP="$x"
+    done
+
+fi
 
 echo wget the ipv4 >> $LOG
 echo "$files" >> $LOG
-
-for x in $files
-do
- PUBLIC_IP="$x"
-done
 
 echo done_ public ip >> $LOG
 echo $PUBLIC_IP >> $LOG
@@ -82,18 +105,19 @@ echo $PUBLIC_IP >> $LOG
 temp_private_ip=`ifconfig eth0 | grep inet | grep -v inet6 | cut -d ":" -f 2 | cut -d " " -f 1`
 PRIVATE_IP=$temp_private_ip
 
-
 for i in `/usr/bin/ruby /opt/get-launch-params.rb`
 do
-  export ${i}
-  echo "export" ${i} >> /home/ubuntu/.bashrc
+    export ${i}
 done
 
-# Copying the product from the controller
-echo "scp -i ${instance_path}/payload/wso2-key root@${CONTROLLER_IP}:/opt/${PRODUCT_NAME}.zip /opt/" >> $LOG
-scp -i ${instance_path}/payload/wso2-key root@${CONTROLLER_IP}:/opt/${PRODUCT_NAME}.zip /opt/
-unzip /opt/${PRODUCT_NAME}.zip -d /opt/
-rm -f /opt/${PRODUCT_NAME}.zip
+
+if [[ ! -d /opt/${PRODUCT_NAME} ]]; then
+    # Copying the product from the controller
+    echo "scp -i ${instance_path}/payload/wso2-key root@${CONTROLLER_IP}:/opt/${PRODUCT_NAME}.zip /opt/" >> $LOG
+    scp -i ${instance_path}/payload/wso2-key root@${CONTROLLER_IP}:/opt/${PRODUCT_NAME}.zip /opt/
+    unzip /opt/${PRODUCT_NAME}.zip -d /opt/
+    rm -f /opt/${PRODUCT_NAME}.zip
+fi
 
 if [ "$ADMIN_USERNAME" = "" ]; then
 	echo Launching with default admin username >> $LOG
@@ -121,3 +145,8 @@ fi
 echo "Starting carbon server ..." >> $LOG
 nohup /opt/${PRODUCT_NAME}/bin/wso2server.sh & >> $LOG
 echo "Carbon server started" >> $LOG
+sleep 1
+if [[ "$(pidof java)" ]]; then
+    crontab -r
+fi
+
